+++ fd-restaurant-service/src/main/java/com/innowise/restaurantservice/service/impl/RestaurantServiceImpl.java
@@ -6,10 +6,9 @@ import com.innowise.restaurantservice.model.Restaurant;
 import com.innowise.restaurantservice.repository.RestaurantRepository;
 import com.innowise.restaurantservice.service.RestaurantService;
 import jakarta.persistence.EntityNotFoundException;
-import java.util.List;
-import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -22,30 +21,17 @@ public class RestaurantServiceImpl implements RestaurantService {
 
   @Override
   @Transactional(readOnly = true)
-  public Restaurant getRestaurant(Long id) {
-    return restaurantRepository.findById(id)
+  public RestaurantDto getRestaurant(Long id) {
+    Restaurant restaurant = restaurantRepository.findById(id)
         .orElseThrow(EntityNotFoundException::new);
+    return restaurantMapper.toDto(restaurant);
   }
 
   @Override
   @Transactional(readOnly = true)
-  public RestaurantDto getRestaurantDto(Long id) {
-    return restaurantMapper.toDto(getRestaurant(id));
-  }
-
-  @Override
-  @Transactional(readOnly = true)
-  public List<Restaurant> getRestaurantList(Integer pageNumber, Integer pageSize) {
-    return restaurantRepository.findAll(PageRequest.of(pageNumber, pageSize))
-        .getContent();
-  }
-
-  @Override
-  @Transactional(readOnly = true)
-  public List<RestaurantDto> getRestaurantDtoList(Integer pageNumber, Integer pageSize) {
-    return getRestaurantList(pageNumber, pageSize).stream()
-        .map(restaurantMapper::toDto)
-        .collect(Collectors.toList());
+  public Page<RestaurantDto> getRestaurantList(Pageable pageable) {
+    return restaurantRepository.findAll(pageable)
+        .map(restaurantMapper::toDto);
   }
 
   @Override
@@ -58,9 +44,11 @@ public class RestaurantServiceImpl implements RestaurantService {
   @Override
   @Transactional
   public void updateRestaurant(Long id, RestaurantDto restaurantDto) {
-    getRestaurant(id);
-    Restaurant convertedRestaurant = restaurantMapper.toEntity(restaurantDto);
+    if (!existsById(id)) {
+      throw new EntityNotFoundException();
+    }
 
+    Restaurant convertedRestaurant = restaurantMapper.toEntity(restaurantDto);
     convertedRestaurant.setId(id);
 
     restaurantRepository.save(convertedRestaurant);
@@ -69,7 +57,17 @@ public class RestaurantServiceImpl implements RestaurantService {
   @Override
   @Transactional
   public void deleteRestaurant(Long id) {
-    getRestaurant(id);
+    if (!existsById(id)) {
+      throw new EntityNotFoundException();
+    }
+
     restaurantRepository.deleteById(id);
+  }
+
+  @Override
+  @Transactional(readOnly = true)
+  public boolean existsById(Long id) {
+    return restaurantRepository.findById(id)
+        .isPresent();
   }
 }
